+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # PwnBox - v2.7.1
-Weekly update.
+
+- Disabled `arm` builds for now since they add so much build time.
+- Added `openvpn` to installed tools.